Fixed code climate issues
